fix not logging in error after registration
not calling login function after register's success callback results in not setting user role. This causes a bug where when a new user is registered(user re-directed to dashboard), if user create a new post there won't be a new post on screen unless user manually refreshes, this is because the user role is not set unless login function is called.